% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_ab.R
\name{BoundaryAB}
\alias{BoundaryAB}
\title{Find Boundary Alpha Beta}
\usage{
BoundaryAB(mu, tol = 1e-06)
}
\arguments{
\item{mu}{Mean of heterogeneity parameter pi.}

\item{tol}{Small tolerance to ensure alpha, beta > 1.}
}
\value{
Numeric vector containing alpha and beta.
}
\description{
Find Boundary Alpha Beta
}
