% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC.R
\name{CalcChi2}
\alias{CalcChi2}
\title{Chi Squared Statistic}
\usage{
CalcChi2(mu_hat, mu, se2)
}
\arguments{
\item{mu_hat}{Estimated mu.}

\item{mu}{True mu.}

\item{se2}{Squared standard error of mu_hat.}
}
\value{
Numeric chi2 statistic.
}
\description{
Chi Squared Statistic
}
