% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bound.R
\name{CheckBoundaryRegion}
\alias{CheckBoundaryRegion}
\title{Check Boundary Region.}
\usage{
CheckBoundaryRegion(
  ab_to_pval_map,
  search_results,
  lower_bound = FALSE,
  mu_extra_steps = 10,
  nu_extra_steps = 10,
  step_size = 2e-04,
  t1e = 0.05
)
}
\arguments{
\item{ab_to_pval_map}{Function that accepts a vector contaiing alpha and
beta then returns a p-value.}

\item{search_results}{Current search results.}

\item{lower_bound}{Logical, searches for the lower bound?}

\item{mu_extra_steps}{Number of steps for mu to check beyond boundary.}

\item{nu_extra_steps}{Number of steps for nu to check beyond boundary.}

\item{step_size}{Distance between successive estimates of mu.}

\item{t1e}{Type I error level.}
}
\value{
Data.frame containing (alpha, beta) values in the check region
where the p-value exceeded the type I error.
}
\description{
Check a grid of mu and nu values in the region of the final value of mu
from the search along the boundary.
}
