% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_conf_int.R
\name{ExactConfInt}
\alias{ExactConfInt}
\title{Outputs the confidence interval.}
\usage{
ExactConfInt(
  size_1,
  events_1,
  size_2,
  events_2,
  reps = 500,
  t1e = 0.05,
  mu0 = NULL,
  step_size = 2e-04,
  maxit = 100
)
}
\arguments{
\item{size_1}{Size of study 1.}

\item{events_1}{Events in study 1.}

\item{size_2}{Size of study 2.}

\item{events_2}{Events in study 2.}

\item{reps}{MC replicates for p-value estimation.}

\item{t1e}{Type I error level.}

\item{mu0}{Optional initial value for mu.}

\item{step_size}{Distance between successive estimates of mu.}

\item{maxit}{Maximum number of iterations to perform.}
}
\value{
Data.frame containing the confidence interval \code{lower}
and \code{upper} bound.
}
\description{
Outputs the confidence interval.
}
\examples{
set.seed(2013)
data <- GenData(
  total_studies = 10,
  n1 = 100,
  n2 = 100,
  alpha2 = 10,
  beta2 = 10
)
# Note: use more high `reps` and smaller `step_size` for more accurate results.
ExactConfInt(
  size_1 = data$size_1,
  events_1 = data$events_1,
  size_2 = data$size_2,
  events_2 = data$events_2,
  reps = 50,
  step_size = 0.02
)
}
