% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bound.R
\name{ExtendBoundary}
\alias{ExtendBoundary}
\title{Extend Boundary}
\usage{
ExtendBoundary(
  mu_to_pval_map,
  search_results,
  keep_history = FALSE,
  lower_bound = FALSE,
  step_size = 2e-04
)
}
\arguments{
\item{mu_to_pval_map}{Function that accepts a value of mu and
returns a p-value.}

\item{search_results}{Current search results.}

\item{keep_history}{Keep search history?}

\item{lower_bound}{Logical, searches for the lower bound?}

\item{step_size}{Distance between successive estimates of mu.}
}
\value{
List containing the current 'bound', the 'mc' results,
and the 'pval' associated with the current bound.
}
\description{
Extends the retention region by moving mu along the boundary
by a single step.
}
