% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bound.R
\name{FindBound}
\alias{FindBound}
\title{Bounds for Mu.}
\usage{
FindBound(
  size_1,
  events_1,
  size_2,
  events_2,
  lower_bound = TRUE,
  reps = 500,
  t1e = 0.05,
  mu0 = NULL,
  step_size = 2e-04,
  maxit = 100,
  keep_history = FALSE,
  mu_num_extra_steps = 10,
  nu_num_extra_steps = 10
)
}
\arguments{
\item{size_1}{Size of study 1.}

\item{events_1}{Events in study 1.}

\item{size_2}{Size of study 2.}

\item{events_2}{Events in study 2.}

\item{lower_bound}{Logical, searches for the lower bound?}

\item{reps}{MC replicates for p-value estimation.}

\item{t1e}{Type I error level.}

\item{mu0}{Optional initial value for mu.}

\item{step_size}{Distance between successive estimates of mu.}

\item{maxit}{Maximum number of iterations to perform.}

\item{keep_history}{Keep search history?}

\item{mu_num_extra_steps}{Number of steps for mu to check beyond boundary.}

\item{nu_num_extra_steps}{Number of steps for nu to check beyond boundary.}
}
\value{
List containing the \code{search_results} and the output of the
\code{boundary_region_check}.
}
\description{
Finds the lower/upper boundary for mu via grid search.
}
\examples{
set.seed(2013)
data <- GenData(
  total_studies = 10,
  n1 = 100,
  n2 = 100,
  alpha2 = 10,
  beta2 = 10
)
# Note: use more `reps` and smaller `step_size` for more accurate results.

# Find lower bound.
FindBound(
  size_1 = data$size_1,
  events_1 = data$events_1,
  size_2 = data$size_2,
  events_2 = data$events_2,
  reps = 50,
  lower_bound = TRUE,  # Set to FALSE for upper bound.
  step_size = 0.02
)
}
