% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{GenData}
\alias{GenData}
\title{Generate Data}
\usage{
GenData(total_studies, n1, n2, alpha1 = 1.44, rate1 = 0.01, alpha2, beta2)
}
\arguments{
\item{total_studies}{Total number of studies.}

\item{n1}{Sample size in group 1.}

\item{n2}{Sample size in group 2.}

\item{alpha1}{Shape parameters for group 1.}

\item{rate1}{Baseline event rate in group 1.}

\item{alpha2}{First shape parameter for beta distribution,
lambda_1/(lambda_1+lambda_2) ~ B(alpha2, beta2).}

\item{beta2}{Second shape parameter for beta distribution,
lambda_1/(lambda_1+lambda_2) ~ B(alpha2, beta2).}
}
\description{
Generate Data
}
