% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bound.R
\name{InitMu}
\alias{InitMu}
\title{Initialize Mu}
\usage{
InitMu(
  data,
  mu_to_pval_map,
  lower_bound = FALSE,
  mu0 = NULL,
  step_size = 2e-04,
  t1e = 0.05
)
}
\arguments{
\item{data}{Study sizes and event counts.}

\item{mu_to_pval_map}{Function that accepts a value of mu and
returns a p-value.}

\item{lower_bound}{Logical, searches for the lower bound?}

\item{mu0}{Optional initial value for mu.}

\item{step_size}{Distance between successive estimates of mu.}

\item{t1e}{Type I error level.}
}
\value{
List containing the current 'bound', the 'mc' results,
and the 'pval' associated with the current bound.
}
\description{
Initializes the search for the boundary mu. If \code{mu0} is provided, directly
assesses the p-value. If not, a moment estimate of mu is obtained, and the
search is initialized from the boundary of the confidence interval for the
moment estimator.
}
