% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_estimator.R
\name{MomentEst}
\alias{MomentEst}
\title{Moment Estimation}
\usage{
MomentEst(size_1, events_1, size_2, events_2, study = NULL)
}
\arguments{
\item{size_1}{Size of study 1.}

\item{events_1}{Events in study 1.}

\item{size_2}{Size of study 2.}

\item{events_2}{Events in study 2.}

\item{study}{Optional study identifier.}
}
\value{
List containing:
\itemize{
\item \code{mu}, estimated first moment.
\item \code{mu_se2}, square standard error of estimated first moment.
\item \code{mu_cc}, estimated first moment with continuity correction.
\item \code{mu2}, estimated second moment.
\item \code{nu}, estimate of nu.
}
}
\description{
Moment Estimation
}
