% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_estimator.R
\name{PrepData}
\alias{PrepData}
\title{Prepare Data for Simulation}
\usage{
PrepData(size_1, events_1, size_2, events_2, study)
}
\arguments{
\item{size_1}{Size of study 1.}

\item{events_1}{Events in study 1.}

\item{size_2}{Size of study 2.}

\item{events_2}{Events in study 2.}

\item{study}{Optional study identifier.}
}
\value{
Data.frame containing
\itemize{
\item \code{study} identifier.
\item \code{events} down-sampled number of events.
\item \code{total_events} total number of events.
\item \code{weight} based on the hypergeometric distribution.
\item \code{n1} and \code{n2}, original study sizes.
}
}
\description{
Prepare Data for Simulation
}
