% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC.R
\name{RunMC}
\alias{RunMC}
\title{Run Monte Carlo Simulation.}
\usage{
RunMC(size_1, events_1, size_2, events_2, study = NULL, reps, alpha, beta)
}
\arguments{
\item{size_1}{Size of study 1.}

\item{events_1}{Events in study 1.}

\item{size_2}{Size of study 2.}

\item{events_2}{Events in study 2.}

\item{study}{Optional study identifier.}

\item{reps}{Replications.}

\item{alpha}{First shape parameter for beta distribution,
lambda_1/(lambda_1+lambda_2) ~ B(alpha, beta).}

\item{beta}{Second shape parameter for beta distribution,
lambda_1/(lambda_1+lambda_2) ~ B(alpha, beta).}
}
\value{
Numeric vector containing:
\itemize{
\item \code{alpha} and \code{beta} of interest.
\item Mean \code{mu} and variance \code{nu} corresponding to \code{alpha} and \code{beta}.
\item P-values based on the normalized and un-normalized test statistics,
\code{p_val_norm} and \code{p_val_unnorm}.
}
}
\description{
Estimates p-values assessing the null hypothesis that the supplied \deqn{\alpha} and
\deqn{\beta} are the generative parameters for the observed event counts.
}
