% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_bound.R
\name{UpperBound}
\alias{UpperBound}
\title{Upper Bound for Mu.}
\usage{
UpperBound(
  size_1,
  events_1,
  size_2,
  events_2,
  reps = 500,
  t1e = 0.05,
  step_size = 2e-04,
  maxit = 100,
  mu0 = NULL,
  keep_history = FALSE
)
}
\arguments{
\item{size_1}{Size of study 1.}

\item{events_1}{Events in study 1.}

\item{size_2}{Size of study 2.}

\item{events_2}{Events in study 2.}

\item{reps}{MC replicates for p-value estimation.}

\item{t1e}{Type I error level.}

\item{step_size}{Distance between successive estimates of mu.}

\item{maxit}{Maximum number of iterations to perform.}

\item{mu0}{Optional initial value for mu.}

\item{keep_history}{Keep search history?}
}
\value{
Data.frame containing the results of \code{\link{RunMC}}
for each \eqn{\mu} in the search interval.
}
\description{
Finds the upper boundary for mu via grid search.
}
\examples{
set.seed(2013)
data <- GenData(
  total_studies = 10,
  n1 = 100,
  n2 = 100,
  alpha2 = 10,
  beta2 = 10
)
# Note: use more high `reps` and smaller `step_size` for more accurate results.
UpperBound(
  size_1 = data$size_1,
  events_1 = data$events_1,
  size_2 = data$size_2,
  events_2 = data$events_2,
  reps = 50,
  step_size = 0.02
)
}
